Make extension modal use new SignWaitingScreenModal
