+++ src/UI/SignTransactionsModals/components/SignWaitingScreenModal/index.tsx
@@ -1,0 +1,1 @@
+export * from './SignWaitingScreenModal';
