+++ src/UI/SignTransactionsModals/components/SignWaitingScreenModal/signWaitingScreenModal.styles.scss
