+++ src/UI/SignTransactionsModals/components/index.tsx
@@ -2,3 +2,4 @@ export * from './ConfirmationScreen';
 export * from './DeviceConfirmationScreen';
 export * from './confirmationScreen.types';
 export * from './TransactionStatusToast';
+export * from './SignWaitingScreenModal';
